requirements.txt
# Core Legal Analysis Dependencies
python-docx==0.8.11

# AI Studio Integration
google-genai>=1.33.0

# NLP Analysis Dependencies (for enhanced legal workflow)
spacy>=3.8.0
textblob>=0.19.0
nltk>=3.9.1

# Web Scraping and HTTP (for court forms downloader)
requests>=2.32.0
beautifulsoup4>=4.12.3
lxml>=5.3.0

# Data Processing and Analysis
pandas>=2.2.0
numpy>=2.1.0

# Configuration and Utilities
pyyaml>=6.0.2
python-dateutil>=2.9.0

# ML and Text Analysis
scikit-learn>=1.7.0
textstat>=0.7.0